+++ CHANGELOG.md
@@ -86,11 +86,12 @@ failure.
 If you are a plugin developer upgrading an existing plugin you will have to
 manually add support for `--failfast` passthrough to your `runtests.py`.
 
-
 #### Misc Changes
 
 Removes the undocumented `collapsed_multisave` tag from the `pathways` templatetag
 library.
+
+Adds a setting `OPAL_FAVICON_PATH` to specify the application Favicon to use.
 
 ### 0.9.1 (Minor Release)